# Conflicts:
#	.idea/runConfigurations/Aper1_Calibration.xml
#	.idea/runConfigurations/Aper1_Myo.xml
#	.idea/runConfigurations/Aper1_R2_PureLogic_6_mm.xml
#	.idea/runConfigurations/Aper2_ECG.xml
#	.idea/runConfigurations/ConverterAper.xml
#	.idea/runConfigurations/Kleiber_Myo.xml
#	.idea/runConfigurations/Nmis.xml
#	.idea/runConfigurations/NmisRsce.xml
#	.idea/runConfigurations/Prv.xml
#	.idea/runConfigurations/Rcm.xml
#	.idea/runConfigurations/RcmCalibration.xml
#	.idea/runConfigurations/SunTechNIBPTest_Aper1.xml
#	.idea/runConfigurations/SunTechNIBP_Aper1.xml
#	.idea/runConfigurations/SunTechNIBP_Aper2.xml
#	src/main/java/com/ak/fx/scene/AxisXController.java

+++ Util/src/main/java/com/ak/util/Clean.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.lang.ref.Cleaner;
 import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
+import java.nio.file.LinkOption;
 import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.logging.Level;
@@ -18,15 +19,30 @@ public enum Clean {
     Arrays.stream(toClean).forEach(Cleaner.Cleanable::clean);
   }
 
-  public static void clean(@Nonnull Path path) {
-    Logger.getLogger(Clean.class.getName()).log(Level.INFO, () -> "Clean directory %s".formatted(path));
-    try (DirectoryStream<Path> ds = Files.newDirectoryStream(path, Files::isRegularFile)) {
-      for (Path file : ds) {
-        Files.deleteIfExists(file);
-      }
+  public static void clean(@Nonnull Path root) {
+    Logger.getLogger(Clean.class.getName()).log(Level.INFO, () -> "Clean directory %s".formatted(root));
+    cleanRecursive(root);
+    try {
+      Files.deleteIfExists(root);
     }
-    catch (IOException ex) {
-      Logger.getLogger(Clean.class.getName()).log(Level.WARNING, path.toString(), ex.getMessage());
+    catch (IOException e) {
+      Logger.getLogger(Clean.class.getName()).log(Level.WARNING, root.toString(), e.getMessage());
+    }
+  }
+
+  private static void cleanRecursive(@Nonnull Path path) {
+    if (Files.exists(path, LinkOption.NOFOLLOW_LINKS)) {
+      try (DirectoryStream<Path> ds = Files.newDirectoryStream(path)) {
+        for (Path file : ds) {
+          if (Files.isDirectory(file, LinkOption.NOFOLLOW_LINKS)) {
+            cleanRecursive(file);
+          }
+          Files.deleteIfExists(file);
+        }
+      }
+      catch (IOException e) {
+        Logger.getLogger(Clean.class.getName()).log(Level.WARNING, path.toString(), e.getMessage());
+      }
     }
   }
 }
